pyproject.toml
[project]
name = "mirrormuse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.26.0",
    "aws-profile-manager==0.7.3",
    "beautifulsoup4>=4.13.4",
    "bitsandbytes>=0.46.1",
    "boto3>=1.39.11",
    "chromedriver-autoinstaller>=0.6.4",
    "click>=8.1.3",
    "datasets==3.0.1",
    "kubernetes==30.1.0",
    "langchain-community==0.2.11",
    "langchain-openai>=0.1.25",
    "loguru>=0.7.3",
    "openai==1.64.0",
    "opik>=1.8.6",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.12.1",
    "s3fs>=0.4.2",
    "sagemaker>=2.232.2",
    "sagemaker-huggingface-inference-toolkit==2.4.0",
    "selenium==4.21.0",
    "sentence-transformers==3.0.0",
    "streamlit>=1.47.0",
    "vllm>=0.6.1",
    "xformers>=0.0.27.post2",
    "zenml[server]==0.74.0",
]
